220704_commit

+++ src/main/java/com/kaveloper/portfolio/dto/BoardListResponseDTO.java
@@ -1,12 +1,10 @@
 package com.kaveloper.portfolio.dto;
 
-import com.kaveloper.portfolio.entity.UploadFile;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import java.time.LocalDateTime;
-import java.util.List;
 
 @Builder
 @AllArgsConstructor
@@ -32,9 +30,9 @@ public class BoardListResponseDTO {
     // 24시간 이후일 경우 연도-날짜 순으로 표시 ex) 2022/06/22
     private long differentTime;
 
-    private int count; // 조회 수
+    private long count; // 조회 수
 
-    private int replyCount; // 댓글 수
+    private long replyCount; // 댓글 수
 
     private int nid; // 공지 사항 여부
